primary refactor work
