--- dspot-prettifier/src/main/java/eu/stamp_project/prettifier/context2name/C2N/python/README.md
@@ -1,15 +1,0 @@
-## Context2Name
-
-### Prepare dependence
-pip3 install bottleneck numpy keras tensorflow
-
-### Process the corpus
-run `PGA.java`
-run `C2N.java`
-
-### Training
-prepare `training.csv`, `validation.csv` and `config.json`
-`python3 c2n_train.py`
-
-### Applying
-run `C2N.java`
