+++ dspot-prettifier/src/main/resources/context2name/demo.txt
@@ -1,0 +1,1 @@
+demo/Demo.java
