+++ dspot-prettifier/src/main/resources/context2name/model/config.json
@@ -1,0 +1,1 @@
+{"UNKNOWN_TOKEN": "**UNK**", "PAD_TOKEN": "**PAD**", "INPUT_VOCAB_SIZE": 4096, "OUTPUT_VOCAB_SIZE": 60000, "N_NEIGHBORS": 10, "KTH_COMMON": 1, "SEQ_LEN": 5, "HIDDEN_LAYER_SIZE": 80, "HIDDEN_LAYER_SIZE2": 350, "BATCH_SIZE": 32, "NUM_EPOCHS": 1, "ACCURACY": 0.8, "ACCURACY2": 0.9, "PLATEAU_LEN": 1, "CHUNK_SIZE1": 2500, "CHUNK_SIZE2": 2000, "PROCESSED_FILE": "vocab.pkl", "TRAINING_FILE": "training.csv", "VALIDATION_FILE": "validation.csv", "MODEL_FILE": "model.h5", "CONFIG_FILE": "config.json"}
